+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This project uses a version scheme based on the Cadenza main version in the form
 ## Unreleased
 ### Added
 - `layout` option for `CadenzaClient#show`
+- `CadenzaClient#batchEditGeometry`
+- `CadenzaClient#batchCreateGeometry`
+- New Sandbox actions for batch editing
+
+### Changed
+- Improved the options type for geometry edit functions
 
 ## 10.4.0 - 2025-06-24